add environment types; move IActiveUser to interfaces/; update DB connection module name
