+++ src/db/connectDB.ts
@@ -1,10 +1,10 @@
 import mongoose from 'mongoose'
 import { timestamp } from '../modules/timestamp.js'
 
-const uri = process.env.MONGO_URI || ''
-
 export async function connectDB() {
 	try {
+		const uri = process.env.MONGO_URI || ''
+
 		mongoose.set('strictQuery', false) // TODO move to options object when standardised
 		await mongoose.connect(uri)
 		console.log(timestamp(), 'Mongo connection made')
