+++ src/types/environment.d.ts
@@ -1,0 +1,12 @@
+export {}
+
+declare global {
+	namespace NodeJS {
+		interface ProcessEnv {
+			MONGO_URI: string
+			OPENCAGE_API_TOKEN: string
+			SESSION_SECRET: string
+			NODE_ENV: 'development' | 'production'
+		}
+	}
+}
